requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
sqlalchemy
psycopg[binary]
redis
boto3
opensearch-py
celery
python-multipart
passlib
bcrypt==3.2.2
pyjwt
openai
pdfminer.six
python-docx
alembic
